+++ src/blocks/team/styling.js
@@ -8,6 +8,7 @@ import generateCSSUnit from "../../../dist/blocks/uagb-controls/generateCSSUnit"
 function styling( props ) {
 
 	const {
+		classMigrate,
 		tag,
 		align,
 		titleColor,
@@ -207,6 +208,9 @@ function styling( props ) {
 
 	var styling_css = ""
 	var id = `#uagb-team-${ props.clientId }`
+	if ( classMigrate ) {
+		id = `.uagb-block-${ props.clientId }`
+	}
 
 	styling_css = generateCSS( selectors, id )
